third_party/blink/web_tests/fast/dom/remove-children-notification-order-expected.txt

CONSOLE ERROR: A descendant of a <select> does not follow the content model.
PASS

PASS
